feat(data): add ProPILE prompt templates and Enron PII sample
The ProPILE paper [1] proposes using known PII to probe LLMs for
memorized personal information. This adds the necessary data files
to support that methodology.

I chose the Enron email corpus as the default PII source because
it was released as public domain by the Federal Energy Regulatory
Commission (FERC) during the 2003 investigation [2], has been widely
used in ML research [3], and is likely present in most LLM training
corpora.

The enron_pii.jsonl file contains a small sample of PII records.
For the full corpus or additional records, see the CMU mirror [4]
or Kaggle [5].

[1]: https://arxiv.org/abs/2307.01881
[2]: https://www.loc.gov/item/2018487913/
[3]: https://en.wikipedia.org/wiki/Enron_Corpus
[4]: https://www.cs.cmu.edu/~enron/
[5]: https://www.kaggle.com/datasets/wcukierski/enron-email-dataset